Небольшие доделки в наземке В основном коде откалиброван акселерометр с MPU

+++ src-ground/src/package_struct.h
@@ -21,10 +21,10 @@ typedef struct
 
 	uint16_t number;	//the number of packet
 
-	uint16_t pressure;	//pressure data
+	int32_t pressure;	//pressure data
 
 	uint16_t temp_ds18b20;
-	uint16_t temp_bmp280;
+	int32_t temp_bmp280;
 
 	uint16_t aXYZ[3];	//accelerometer raw data
 	uint16_t gXYZ[3];	//gyro raw data
